requirements_wmh.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
cachetools==5.3.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
charset-normalizer==3.1.0
click==8.1.3
cycler==0.11.0
docker-pycreds==0.4.0
flatbuffers==2.0.7
fonttools==4.38.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.4.1
joblib==1.2.0
Keras==2.0.7
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pathtools==0.1.2
Pillow==9.5.0
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-wavelets
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.19.1
setproctitle==1.3.2
SimpleITK==2.2.1
six==1.16.0
sklearn==0.0.post4
smmap==5.0.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
Theano==1.0.5
threadpoolctl==3.1.0
torch==1.10.0
torchaudio==0.13.1
torchvision==0.11.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.14.2
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
